# Validating, __init__-based assoc.