--- a/src/attr/_funcs.py
+++ b/src/attr/_funcs.py
@@ -165,14 +165,22 @@ def assoc(inst, **changes):
     :raise attr.exceptions.NotAnAttrsClassError: If *cls* is not an ``attrs``
         class.
     """
-    new = copy.copy(inst)
-    attrs = fields(inst.__class__)
-    for k, v in iteritems(changes):
-        a = getattr(attrs, k, NOTHING)
-        if a is NOTHING:
-            raise AttrsAttributeNotFoundError(
-                "{k} is not an attrs attribute on {cl}."
-                .format(k=k, cl=new.__class__)
-            )
-        _obj_setattr(new, k, v)
-    return new
+    cls = inst.__class__
+    for a in fields(cls):
+        attr_name = a.name  # To deal with private attributes.
+        if attr_name[0] == "_":
+            init_name = attr_name[1:]
+            if attr_name not in changes:
+                changes[init_name] = getattr(inst, attr_name)
+            else:
+                # attr_name is in changes, it needs to be translated.
+                changes[init_name] = changes.pop(attr_name)
+        else:
+            if attr_name not in changes:
+                changes[attr_name] = getattr(inst, attr_name)
+    try:
+        return cls(**changes)
+    except TypeError as exc:
+        k = exc.args[0].split("'")[1]
+        raise AttrsAttributeNotFoundError(
+            "{k} is not an attrs attribute on {cl}.".format(k=k, cl=cls))