@hynek 2016-12-19
Nah we don’t have to support both.  I had a slightly different approach in mind (which I honestly forgot :)) which would’ve given it us for free.  Other than that I don’t have much to add.  The performance gains look awesome!

We can always add support for underscore names if enough people complain.  But removing is much harder.