@Tinche 2016-12-19
Just to clarify, this implementation supports underscore names, i.e. the setter syntax.