@hynek 2016-12-19
Oh, so it doesn’t support the `__init__` syntax?