@hynek 2016-12-19
No it’s alright.  Backward compatibility FTW.  Please finish up. :)