@hynek 2016-12-21
I’ll have to meditate on the backward-incompatible part.

To me, this is a bug fix because I think everyone would expect it to behave this way.  The point about `__init__` is true tho.  Technically we’d have to announce it now and wait for a year though.

Hmmmm.