@hynek 2016-12-21
In any case:

- we’ll have to add a note to the assoc docstring about the __init__ situation
- you have to stop using Markdown in rST :)
- it shouldn’t be called `attrs`-style __init__ but described what it actually means instead.  It looks like the init has to be literally written by attrs like this.