@glyph 2016-12-21
The usual guidance is in Twisted when people start fretting about changed behavior is "new names are cheaper than you think".  Instead of `assoc`, have `assoc2`; deprecate `assoc`.  Then, come up with a good name for `assoc2`.

FWIW I have always hated `assoc` as a name; it's an appeal to lisp-nerdery and nobody knows how to pronounce it.  I think `clone` expresses this functionality in terms that Python peeps are more likely to use.  But, even if it's just `assoc` in a different namespace, that would be a way to make a clean break.