@glyph 2016-12-21
( I should also note, though, that this is a _good_ change.  I was just reviewing a PR on https://github.com/rackerlabs/otter where I was lamenting the odd behavior of `assoc` and its inconsistencies with `__init__`. )