@Tinche 2017-01-06
I'm still with the inlaws for the holidays and away from my home computer. I've been thinking of the name, though. I think Glyph is right and we should leave assoc as is, deprecate it, and think of a better name for this implementation.

What's actually happening is:
* a shallow copy is being made
* with optional changes incorporated

'Copy' or 'clone' are okay but they don't really mesh well with the fact the original object is being changed. In fact, they kinda imply the object is not being changed at all. In my opinion.

How about 'recons', as in 'reconstruct' or 'reconstitute'? We're breaking the instance into parts (asdict), then putting it back together, changed.

    >>> attr.recons(obj, x=1)

(`attr.reconsitute` is too long I think.)

Another idea is `attr.reinit`. Init evokes the initializer.