@glyph 2017-01-06
All of the various derivatives of the word "change" - "modify", "mutate", etc, imply the thing itself is changing, as opposed to a new thing coming into existence.

But wait - in that sentence I used the word "derivative" to describe exactly that sort of thing!  This might be too cutesy given its overlap with other terms, especially math terms, but `derive` might be the word we're looking for.  If you have A, and you want A', where A' is like A but with some changes, you could say A' is `derive`d from A.

I could see `attr.derive(obj, x=1)` being idiomatic.

Thoughts?