@hynek 2017-01-06
I agree that derive is the first name that I could sort of live with but I'm not happy with it.

There *has* to be *some* term that isn't some obscure math/FP lingo and that expresses precisely what we need!?

I really don't want it to be unclear what it's about.