@glyph 2017-01-06
> There has to be some term that isn't some obscure math/FP lingo and that expresses precisely what we need!?

["convolve"](http://www.dictionary.com/browse/convolve?s=t)? :)