@hynek 2017-01-06
OK one more rule: I don’t have to look it up in a dictionary. ಠ_ಠ