@hynek 2017-01-06
`evolve` wouldn't be too bad but also kind of obscure. :-/