@hynek 2017-01-06
FTR, pyrsistent seems to have `transform` which doesn’t seem too bad?