@glyph 2017-01-06
> FTR, pyrsistent seems to have transform which doesn’t seem too bad?

Except for the fact that "transform" also means "change" in the "mutate the existing object" sort of way.