@glyph 2017-01-06
> evolve wouldn't be too bad but also kind of obscure. :-/

Actually this might be the best bet. Consider that even the average 7-year-old knows that Pokémon have "evolutions".*

*: except of course what pokémon _actually_ have are metamorphoses, which are destructive changes, but as a _real_ species "evolves" it involves creating lots of new instances