@hynek 2017-01-06
I was rather thinking about evolution which usually happens over generations.  Humans evolved from ape.  There was no ape who suddenly liked computers.