@hynek 2017-01-11
So I believe we’ve agreed on deprecating assoc as it is and implementing a new `evolve()`, correct?