@Tinche 2017-01-11
Alright, so we're settled on evolve? I prefer derive, but the boss has the final say of course :)

I'll put together a new PR soon-ish. Are we deprecating assoc though?