@glyph 2017-01-12
> So I believe we’ve agreed on deprecating assoc as it is and implementing a new evolve(), correct?

Yay! [Evolvulation!](https://github.com/twisted/twisted/blob/e38cc25a67747899c6984d6ebaa8d3d134799415/docs/historic/Quotes/Twisted-8.0#L70-L74)