@hynek 2017-01-12
I can live with derive too...I guess I'll have to check a thesaurus or something. :-/