@glyph 2017-01-12
Sorry, I didn't mean that to seem critical of the choice of "evolve", I just found the quote evocative :).

`evolve` is great. If you want to consult a thesaurus, don't let me stop you, but don't go back to `derive` purely on my account.