@hynek 2017-01-12
I like both for different reasons. :|