@Tinche 2017-01-12
Working on this a little.

The docs state: 

> If there will ever be a need to break compatibility, it will be announced in the Changelog and raise a DeprecationWarning for a year (if possible) before it’s finally really broken.

Ok, let's raise a DeprecationWarning, as per the docs.

    import warnings
    warnings.warn("deprecated", DeprecationWarning)

Nothing happens. Pop over to https://docs.python.org/3.6/library/warnings.html:

> By default, Python installs several warning filters, which can be overridden by the command-line options passed to -W and calls to filterwarnings().
>
>   * DeprecationWarning and PendingDeprecationWarning, and ImportWarning are ignored.
> ...
> 
> Changed in version 3.2: DeprecationWarning is now ignored by default in addition to PendingDeprecationWarning.

Doesn't this make DeprecationWarning basically useless? Who runs their Python with -W? Not me. In fact I had to Google how to enable warnings at all (it's python -Wd).