@bluetech 2017-01-13
Regarding the name, IIUC (I have very little experience with the library, so maybe not), this operation is similar to [namedtuple._replace](https://docs.python.org/3.6/library/collections.html#collections.somenamedtuple._replace), in which case `replace` might be appropriate.