@hynek 2017-01-14
# Deprecation Warnings

We should add documentation on them to the backward compatibility page.

# Naming

I’m not a fan of `replace()` because it doesn’t express what it does at all (for namedtuples the copy is implicit because they’re always immutable).  But we went with namedtuple lingo before (asdict) so there’s a precedent and I’m not entirely a fan of the others either. :|

So far we have:

- `replace()`: established lingo, no semantics in name
- `evolve()`: semantics can be guessed but obscure
- `derive()`: i’ve heard people calling subclassing as derivation too so that might be confusing.

I’ve spend some time in a thesaurus and found nothing better.

Anyone more 🎨 ?